requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic==0.34.0
python-dotenv==1.0.0
pydantic==2.5.0
asyncio-mqtt==0.16.1
aiofiles==23.2.0
python-multipart==0.0.6
jinja2==3.1.2